Moved the Dice Context Split the statistics state into two smaller interfaces
