+++ src/Components/Dice/DiceContext.ts
@@ -1,0 +1,6 @@
+import { createContext } from "react";
+import { PoolDice } from "../../Models";
+
+const emptyData: PoolDice[] = [];
+
+export const DiceContext = createContext(emptyData)
