--- src/Components/DiceContext.ts
@@ -1,6 +1,0 @@
-import React from "react";
-import { PoolDice } from "../Models";
-
-const emptyData: PoolDice[] = [];
-
-export const DiceContext = React.createContext(emptyData)
